リファクタリング: SynthesisEngineをモジュール化 (#214)
* synthesis_engine module

* split make_synthesis_engine

* use relative import

* rm noqa

* add comment

* fix format

* add comment

* use snake_case for class module (pep8)

* fix format

* fix check_tts import
