+++ voicevox_engine/dev/synthesis_engine/mock.py
@@ -5,8 +5,8 @@ import numpy as np
 from pyopenjtalk import tts
 from scipy.signal import resample
 
-from voicevox_engine.model import AccentPhrase, AudioQuery
-from voicevox_engine.synthesis_engine import to_flatten_moras
+from ...model import AccentPhrase, AudioQuery
+from ...synthesis_engine.synthesis_engine import to_flatten_moras
 
 
 class SynthesisEngine:
